# openapi.yaml
openapi: 3.1.0
info:
  title: AI City Guide Backend API
  version: 0.1.0
  description: Production API for AI-powered city search, recommendation, auth, favorites, and statistics.
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
paths:
  /health:
    get:
      tags: [health]
      summary: Health check
      security: []
      responses:
        '200':
          description: Service health snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User created and authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags: [auth]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token and issue new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refreshed session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /me:
    get:
      tags: [me]
      summary: Get current user profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
  /search:
    post:
      tags: [search]
      summary: Perform AI-powered city search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Ranked recommendation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '404':
          $ref: '#/components/responses/NoSearchResultsError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /history:
    get:
      tags: [history]
      summary: Get search history for the authenticated user
      responses:
        '200':
          description: Search history entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchHistoryResponse'
  /favorites:
    get:
      tags: [favorites]
      summary: List favorite places
      responses:
        '200':
          description: Favorite places
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FavoriteResponse'
    post:
      tags: [favorites]
      summary: Add a favorite place
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteCreateRequest'
      responses:
        '200':
          description: Favorite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteResponse'
  /favorites/{id}:
    delete:
      tags: [favorites]
      summary: Delete a favorite place
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /statistics:
    get:
      tags: [statistics]
      summary: Daily search statistics for the authenticated user
      responses:
        '200':
          description: Statistics records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchStatisticsResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    AuthenticationError:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NoSearchResultsError:
      description: No places matched the query
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CoordinatesSchema:
      type: object
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
      required: [latitude, longitude]
      additionalProperties: false
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        full_name:
          type: string
          minLength: 2
          maxLength: 255
      required: [email, password, full_name]
      additionalProperties: false
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
      additionalProperties: false
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
      additionalProperties: false
    LogoutRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
      additionalProperties: false
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
      additionalProperties: false
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          const: bearer
        expires_in:
          type: integer
      required: [access_token, refresh_token, token_type, expires_in]
      additionalProperties: false
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [user, admin]
        is_active:
          type: boolean
      required: [id, email, full_name, role, is_active]
      additionalProperties: false
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        tokens:
          $ref: '#/components/schemas/AuthTokens'
      required: [user, tokens]
      additionalProperties: false
    SearchRequest:
      type: object
      properties:
        query:
          type: string
          minLength: 3
          maxLength: 1000
        coordinates:
          $ref: '#/components/schemas/CoordinatesSchema'
        locale:
          type: string
          default: en
      required: [query]
      additionalProperties: false
    SearchIntentSchema:
      type: object
      properties:
        query:
          type: string
        location_text:
          type: [string, 'null']
        coordinates:
          $ref: '#/components/schemas/CoordinatesSchema'
        radius_m:
          type: integer
          minimum: 100
          maximum: 10000
        budget_kzt:
          type: [integer, 'null']
          minimum: 0
          maximum: 1000000
        party_size:
          type: integer
          minimum: 1
          maximum: 50
        cuisine:
          type: [string, 'null']
        place_type:
          type: [string, 'null']
        amenities:
          type: array
          items:
            type: string
        mood:
          type: [string, 'null']
        sort_by:
          type: string
        open_now:
          type: boolean
        min_rating:
          type: number
          minimum: 0
          maximum: 5
        price_category:
          type: [string, 'null']
        requires_parking:
          type: boolean
        requires_quiet:
          type: boolean
        laptop_friendly:
          type: boolean
        romantic:
          type: boolean
      required: [query, radius_m, party_size, amenities, sort_by, open_now, min_rating, requires_parking, requires_quiet, laptop_friendly, romantic]
      additionalProperties: false
    PlaceRecommendationSchema:
      type: object
      properties:
        place_id:
          type: string
        name:
          type: string
        rating:
          type: [number, 'null']
        walking_time:
          type: [integer, 'null']
        pros:
          type: array
          items:
            type: string
        cons:
          type: array
          items:
            type: string
        reason:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        score:
          type: number
        address:
          type: [string, 'null']
        categories:
          type: array
          items:
            type: string
        distance_m:
          type: [integer, 'null']
        price_category:
          type: [string, 'null']
        opening_hours:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        url:
          type: [string, 'null']
      required: [place_id, name, pros, cons, reason, confidence, score, categories]
      additionalProperties: false
    SearchResponse:
      type: object
      properties:
        recommendation:
          $ref: '#/components/schemas/PlaceRecommendationSchema'
        alternatives:
          type: array
          items:
            $ref: '#/components/schemas/PlaceRecommendationSchema'
        intent:
          $ref: '#/components/schemas/SearchIntentSchema'
        source:
          type: string
        generated_at:
          type: string
          format: date-time
      required: [recommendation, alternatives, intent, source, generated_at]
      additionalProperties: false
    FavoriteCreateRequest:
      type: object
      properties:
        place_id:
          type: string
          minLength: 1
          maxLength: 128
        place_name:
          type: string
          minLength: 1
          maxLength: 255
        payload:
          type: object
          additionalProperties: true
        note:
          type: [string, 'null']
          maxLength: 500
      required: [place_id, place_name, payload]
      additionalProperties: false
    FavoriteResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        place_id:
          type: string
        place_name:
          type: string
        payload:
          type: object
          additionalProperties: true
        note:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
      required: [id, place_id, place_name, payload, created_at]
      additionalProperties: false
    SearchHistoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        query:
          type: string
        intent:
          type: object
          additionalProperties: true
        result:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
      required: [id, query, intent, result, created_at]
      additionalProperties: false
    SearchStatisticsResponse:
      type: object
      properties:
        stat_date:
          type: string
          format: date-time
        user_id:
          type: [string, 'null']
          format: uuid
        total_searches:
          type: integer
        successful_searches:
          type: integer
      required: [stat_date, user_id, total_searches, successful_searches]
      additionalProperties: false
    HealthResponse:
      type: object
      properties:
        status:
          type: string
        database:
          type: string
        redis:
          type: string
        external_services:
          type: object
          additionalProperties:
            type: string
      required: [status, database, redis, external_services]
      additionalProperties: false
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: [object, 'null']
          additionalProperties: true
      required: [error, message]
      additionalProperties: false